this is the files inside storefront-child theme
This is just and assignment test. Based on the tasks i've made a little change accordingly (Search AJAX) while i used DataTables which is faster and more simple. I've even added features in the tables (see cities-ajax.js) search is faster, paging, and ordering which is helpful for bulk contents.